fjern unødig mellomlag mellom api og backend-kall
