+++ app/actions/aktivitet.action.server.ts
@@ -3,22 +3,10 @@ import { validationError } from "remix-validated-form";
 import { lagreAktivitet } from "~/models/aktivitet.server";
 import { IRapporteringsperiodeDag } from "~/models/rapporteringsperiode.server";
 
-import { oppdaterAktiviteter } from "./aktivitet.action.utils";
-import { AktivitetType } from "./aktivitettype.utils";
-import type { INetworkResponse } from "./types";
-import { validator } from "./validering.util";
-
-export async function slettAlleAktiviteterForBestemtDag(
-  request: Request,
-  periodeId: string,
-  formdata: FormData,
-): Promise<INetworkResponse> {
-  const dag: IRapporteringsperiodeDag = JSON.parse(String(formdata.get("dag")));
-
-  const oppdatertDag = { ...dag, aktiviteter: [] };
-
-  return await lagreAktivitet(request, periodeId, oppdatertDag);
-}
+import { oppdaterAktiviteter } from "../utils/aktivitet.utils";
+import { AktivitetType } from "../utils/aktivitettype.utils";
+import type { INetworkResponse } from "../utils/types";
+import { validator } from "../utils/validering.util";
 
 export async function validerOgLagreAktivitet(
   request: Request,
