+++ app/utils/aktivitet.utils.ts
