+++ tests/actions/aktivitet.action.utils.test.ts
@@ -2,7 +2,7 @@ import { uuidv7 } from "uuidv7";
 import { describe, expect, it, vi } from "vitest";
 
 import { IRapporteringsperiodeDag } from "~/models/rapporteringsperiode.server";
-import { oppdaterAktiviteter } from "~/utils/aktivitet.action.utils";
+import { oppdaterAktiviteter } from "~/utils/aktivitet.utils";
 import { AktivitetType } from "~/utils/aktivitettype.utils";
 
 vi.mock("uuidv7", () => ({
